+++ app/src/proto/plugins_grpc_web_pb.js
@@ -21,6 +21,8 @@ var prometheus_pb = require('./prometheus_pb.js')
 var elasticsearch_pb = require('./elasticsearch_pb.js')
 
 var jaeger_pb = require('./jaeger_pb.js')
+
+var opsgenie_pb = require('./opsgenie_pb.js')
 const proto = {};
 proto.plugins = require('./plugins_pb.js');