+++ pkg/api/plugins/plugins/plugins.go
@@ -7,6 +7,7 @@ import (
 	clustersProto "github.com/kobsio/kobs/pkg/api/plugins/clusters/proto"
 	"github.com/kobsio/kobs/pkg/api/plugins/elasticsearch"
 	"github.com/kobsio/kobs/pkg/api/plugins/jaeger"
+	"github.com/kobsio/kobs/pkg/api/plugins/opsgenie"
 	pluginsProto "github.com/kobsio/kobs/pkg/api/plugins/plugins/proto"
 	"github.com/kobsio/kobs/pkg/api/plugins/prometheus"
 	"github.com/kobsio/kobs/pkg/config"
@@ -69,10 +70,17 @@ func Register(cfg *config.Config, grpcServer *grpc.Server) error {
 		return err
 	}
 
+	opsgenieInstances, err := opsgenie.Register(cfg.Opsgenie, grpcServer)
+	if err != nil {
+		log.WithError(err).WithFields(logrus.Fields{"plugin": "jaeger"}).Errorf("Failed to register Opsgenie plugin.")
+		return err
+	}
+
 	var plugins []*pluginsProto.PluginShort
 	plugins = append(plugins, prometheusInstances...)
 	plugins = append(plugins, elasticsearchInstances...)
 	plugins = append(plugins, jaegerInstances...)
+	plugins = append(plugins, opsgenieInstances...)
 
 	pluginsProto.RegisterPluginsServer(grpcServer, &Plugins{
 		plugins: plugins,