+++ pkg/api/plugins/plugins/proto/plugins_deepcopy.gen.go
@@ -8,6 +8,7 @@ import (
 	proto "github.com/gogo/protobuf/proto"
 	_ "github.com/kobsio/kobs/pkg/api/plugins/elasticsearch/proto"
 	_ "github.com/kobsio/kobs/pkg/api/plugins/jaeger/proto"
+	_ "github.com/kobsio/kobs/pkg/api/plugins/opsgenie/proto"
 	_ "github.com/kobsio/kobs/pkg/api/plugins/prometheus/proto"
 	math "math"
 )